pyproject.toml
[tool.poetry]
name = "ArcaAutoVep-automatizations"
version = "0.1.0"
description = "ArcaAutoVep - Automatizations System"
authors = ["Joan Canossa <jfcpersonal99@gmail.com>"]
readme = "README.md"
packages = [
    {include = "api", from = "src"},
    {include = "core", from = "src"}
]

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.15.0"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
redis = "^5.0.0"
prometheus-client = "^0.19.0"
python-json-logger = "^2.0.7"
kafka-python = "^2.0.2"
cryptography = "^46.0.3"
reportlab = "^4.4.4"
pandas = "^2.3.3"
openpyxl = "^3.1.5"
google-api-python-client = "^2.187.0"
google-auth-httplib2 = "^0.2.1"
google-auth-oauthlib = "^1.2.3"
debugpy = "^1.8.17"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pre-commit = "^4.3.0"
isort = "^6.0.1"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | __pycache__
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=."
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
